Implemented demo script for minlysreklamen, as well as fixing some bugs

+++ generator/src/java/darkness/generator/api/ScriptBase.java
@@ -106,7 +106,7 @@ public abstract class ScriptBase extends Generator<Void> {
      * @param saturation In the range 0.0..1.0
      * @param brightness In the range 0.0..1.0
      */
-    protected void setHSB(BulbSet bulbSet, int hue, int saturation, int brightness) {
+    protected void setHSB(BulbSet bulbSet, float hue, float saturation, float brightness) {
         bulbSet.setHSB(hue, saturation, brightness);
     }
